+++ python_plus/python_plus/scripts/vem.sh
@@ -372,6 +372,7 @@ pip_install() {
   tmpdir=$VIRTUAL_ENV/tmp
   pkg=$(get_actual_pkg "$pkg")
   pfn=$(get_pkg_wo_version "$pkg")
+  [[ $pfn == "zerobug" ]] && set -x   #debug
   x="-qP"
   [[ -n "$opt_pyver" ]] && x="$x -y$opt_pyver"
   [[ -n "$opt_oever" ]] && x="$x -b$opt_oever"
@@ -510,6 +511,7 @@ pip_install() {
     x="${pkgs//+/.}"
     [[ -z $XPKGS_RE ]] && XPKGS_RE="$x" || XPKGS_RE="$XPKGS_RE|$x"
   fi
+  set +x  #debug
   return $sts
 }